+++ lib/cuckoo/common/integrations/parse_pe.py
@@ -907,7 +907,11 @@ class PortableExecutable(object):
             if not IsPEImage(f.read()):
                 return {}
 
-        pe = pefile.PE(self.file_path)
+        try:
+            pe = pefile.PE(self.file_path)
+        except Exception as e:
+            log.error("PE type not recognised: %s", str(e))
+            return {}
         peresults = {}
         peresults["guest_signers"] = self.get_guest_digital_signers(task_id)
         peresults["digital_signers"] = self.get_digital_signers(pe)