+++ lib/cuckoo/common/thirdparty_detections.py
@@ -10,39 +10,41 @@ def get_crowdstrike_family(proctype, procres):
     """
     :param proctype: process results type
     :param procres: process results
-    :return maldata: list of dicts of yara based family name and actor, or empty list
+    :return maldata: list of dicts of sha256 hash and yara based family, or empty list
     """
     maldata = []
     if proctype == "target":
         yarahits = procres.get("file", {}).get("yara", [])
+        sha256 = procres.get("file", {}).get("sha256", "")
+        families = []
         for yh in yarahits:
             mf = yh.get("meta",{}).get("malware_family", "")
             if mf:
-                maldata.append({"malware_family": mf})
-            ma = yh.get("meta",{}).get("actor", "")
-            if ma:
-                maldata.append({"actor": ma})
+                families.append(mf)
+        maldata.append({sha256: families})
     elif proctype == "CAPE":
-        payloads = procres.get("payload", {})
+        payloads = procres.get("payloads", {})
         for payload in payloads:
             yarahits = payload.get("yara", [])
+            sha256 = payload.get("sha256", "")
+            families = []
             for yh in yarahits:
                 mf = yh.get("meta",{}).get("malware_family", "")
                 if mf:
-                    maldata.append({"malware_family": mf})
-                ma = yh.get("meta",{}).get("actor", "")
-                if ma:
-                    maldata.append({"actor": ma})
+                    families.append(mf)
+            if families:
+                maldata.append({sha256: families})
     elif proctype in ("dropped", "procdump", "procmemory"):
         for data in procres:
             yarahits = data.get("yara", [])
+            sha256 = data.get("sha256", "")
+            families = []
             for yh in yarahits:
                 mf = yh.get("meta",{}).get("malware_family", "")
                 if mf:
-                    maldata.append({"malware_family": mf})
-                ma = yh.get("meta",{}).get("actor", "")
-                if ma:
-                    maldata.append({"actor": ma})
+                    families.append(mf)
+            if families:
+                maldata.append({sha256: families})
 
     return maldata
 
@@ -56,8 +58,8 @@ def get_mandiant_name(identifier):
     ids = {}
 
     try:
-        if identifier.startswith("FE_"):
-            name = re.findall(name_re1, identifier)[0]
+        if identifier.startswith(("FE_", "MTI_")):
+            name = re.findall(name_re1, identifier)[0][1]
             ids[name] = 1
         else:
             return
@@ -166,16 +168,19 @@ def get_mandiant_name(identifier):
     omitlist = set(omitlist)
 
     for key in ids.keys():
-        parts = key.split("_")
-        matches = omitlist.intersection(set(parts))
-        if matches:
-            finparts = [x for x in parts if x not in matches]
-        else:
-            finparts = parts
-        if len(finparts) > 1:
-            name = "_".join(finparts)
-        else:
-            name = finparts[0]
+        try:
+            parts = key.split("_")
+            matches = omitlist.intersection(set(parts))
+            if matches:
+                finparts = [x for x in parts if x not in matches]
+            else:
+                finparts = parts
+            if len(finparts) > 1:
+                name = "_".join(finparts)
+            else:
+                name = finparts[0]
+        except Exception as e:
+            pass
 
     return name
 
@@ -184,35 +189,47 @@ def get_mandiant_family(proctype, procres):
     """
     :param proctype: process results type
     :param procres: process results
-    :return maldata: list of dicts of yara based family name, or empty list
+    :return maldata: list of dicts of sha256 hash and yara based family, or empty list
     """
     maldata = []
     if proctype == "target":
         yarahits = procres.get("file", {}).get("yara", [])
+        sha256 = procres.get("file", {}).get("sha256", "")
+        families = []
         for yh in yarahits:
             identifier = yh.get("name", "")
             if identifier:
                 mf = get_mandiant_name(identifier)
                 if mf:
-                    maldata.append({"malware_family": mf})
+                    families.append(mf)
+        if families:
+            maldata.append({sha256: families})
     elif proctype == "CAPE":
-        payloads = procres.get("payload", {})
+        payloads = procres.get("payloads", {})
         for payload in payloads:
             yarahits = payload.get("yara", [])
+            sha256 = payload.get("sha256", "")
+            families = []
             for yh in yarahits:
                 identifier = yh.get("name", "")
                 if identifier:
                     mf = get_mandiant_name(identifier)
                     if mf:
-                        maldata.append({"malware_family": mf})
+                        families.append(mf)
+            if families:
+                maldata.append({sha256: families})
     elif proctype in ("dropped", "procdump", "procmemory"):
         for data in procres:
             yarahits = data.get("yara", [])
+            sha256 = data.get("sha256", "")
+            families = []
             for yh in yarahits:
                 identifier = yh.get("name", "")
                 if identifier:
                     mf = get_mandiant_name(identifier)
                     if mf:
-                        maldata.append({"malware_family": mf})
+                        families.append(mf)
+            if families:
+                maldata.append({sha256: families})
 
     return maldata