+++ lib/cuckoo/common/web_utils.py
@@ -202,12 +202,15 @@ def top_detections(date_since: datetime = False, results_limit: int = 20) -> dic
             return data
 
     """function that gets detection: count
-    based on: https://gist.github.com/clarkenheim/fa0f9e5400412b6a0f9d
+    Original: https://gist.github.com/clarkenheim/fa0f9e5400412b6a0f9d
+    New: https://stackoverflow.com/a/21509359/1294762
     """
 
     aggregation_command = [
-        {"$match": {"detections": {"$exists": True}}},
-        {"$group": {"_id": "$detections", "total": {"$sum": 1}}},
+        {"$match": {"detections.family": {"$exists": True}}},
+        {"$project": {"_id": 0, "detections.family": 1}},
+        {"$unwind": "$detections"},
+        {"$group": {"_id": "$detections.family", "total": {"$sum": 1}}},
         {"$sort": {"total": -1}},
         {"$addFields": {"family": "$_id"}},
         {"$project": {"_id": 0}},
@@ -220,16 +223,16 @@ def top_detections(date_since: datetime = False, results_limit: int = 20) -> dic
     if repconf.mongodb.enabled:
         data = mongo_aggregate("analysis", aggregation_command)
     elif repconf.elasticsearchdb.enabled:
+        # ToDo update to new format
         q = {
-            "query": {"bool": {"must": [{"exists": {"field": "detections"}}]}},
+            "query": {"bool": {"must": [{"exists": {"field": "detections.family"}}]}},
             "size": 0,
-            "aggs": {"family": {"terms": {"field": "detections.keyword", "size": results_limit}}},
+            "aggs": {"family": {"terms": {"field": "detections.family.keyword", "size": results_limit}}},
         }
 
         if date_since:
             q["query"]["bool"]["must"].append({"range": {"info.started": {"gte": date_since.isoformat()}}})
 
-            print(q)
         res = es.search(index=get_analysis_index(), body=q)
         data = [{"total": r["doc_count"], "family": r["key"]} for r in res["aggregations"]["family"]["buckets"]]
     else:
@@ -830,9 +833,8 @@ def validate_task(tid, status=TASK_REPORTED):
 perform_search_filters = {
     "info": 1,
     "virustotal_summary": 1,
-    "detections": 1,
+    "detections.family": 1,
     "malfamily_tag": 1,
-    "actor": 1,
     "malscore": 1,
     "network.pcap_sha256": 1,
     "mlist_cnt": 1,
@@ -869,8 +871,7 @@ search_term_map = {
     "ip": "network.hosts.ip",
     "signature": "signatures.description",
     "signame": "signatures.name",
-    "detections": "detections",
-    "actor": "actor",
+    "detections": "detections.family",
     "url": "target.url",
     "iconhash": "static.pe.icon_hash",
     "iconfuzzy": "static.pe.icon_fuzzy",